Demented earplug/issue 141

+++ OdinAPI/handler/user.py
@@ -35,7 +35,7 @@ class UserHandler:
         permissionsDictionary['hash'] = record[0]
         return permissionsDictionary
 
-    def addDashUser(self, username, fullName,email, password):
+    def addDashUser(self, username, fullName, email, password):
         """
         Adds a new Dashboard user with the provided information.
 
@@ -48,32 +48,38 @@ class UserHandler:
             full_name: The name of the new dashboboard user.
             email: The email of the new dashboboard user.
             password: The hash of the password for the new dashboboard user.
-            
+
         Returns:
             A JSON containing all the user with the new dashboard user.
         """
+        # Verify inputs are not empty
+        if username == None or fullName == None or email == None or password == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
+
         # Verify password complies with the rules
         if not rulesMatch(password):
             return jsonify(Error="""Password should contain At least 1 upercase letter,
             1 lowecase letter, at least 1 number, at least 1 symbol, and is between 
             10 and 64 characters long."""), 400
 
-        # Hash the password 
+        # Hash the password
         hashedPassword = createHash(password)
 
         dao = UserDAO()
-        res = dao.addDashUser(username, fullName,email, hashedPassword)
+        res = dao.addDashUser(username, fullName, email, hashedPassword)
 
         if res == 'UserError1':
-            return jsonify(Error='Email has been registered.'),400 # Conflict with the current state of the server
+            # Conflict with the current state of the server
+            return jsonify(Error='Email has been registered.'), 400
         elif res == 'UserError2':
-            return jsonify(Error='Username is already taken.'),400
+            return jsonify(Error='Username is already taken.'), 400
         elif res == 'UserError3':
             return jsonify(Error="New user not created"), 400
         else:
             mappedUser = self.mapUserToDict(res)
-            dao.setLoginAttempts(mappedUser['id'],0) #sets login attempts to 0 upon creation.
-            return jsonify(User=mappedUser),201
+            # sets login attempts to 0 upon creation.
+            dao.setLoginAttempts(mappedUser['id'], 0)
+            return jsonify(User=mappedUser), 201
 
     def getAllDashUsers(self):
         """
@@ -91,13 +97,13 @@ class UserHandler:
 
         if userList == None:
             return jsonify(Error='No users found in the system'), 404
-        
+
         mappedUsers = []
         for user in userList:
-            mappedUsers.append(self.mapUserToDict(user)) 
-        
-        return jsonify(Users=mappedUsers),200 #200 == OK
-        
+            mappedUsers.append(self.mapUserToDict(user))
+
+        return jsonify(Users=mappedUsers), 200  # 200 == OK
+
     def getDashUserByID(self, duid):
         """
         Gets a single dashboard user given their ID.
@@ -108,18 +114,20 @@ class UserHandler:
 
         Args:
             duid: The ID of the dashboboard user that needs to be fetched.
-            
+
         Returns:
             A JSON containing all the user with the given ID.
         """
+        if duid == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
         dao = UserDAO()
         fetchedUser = dao.getDashUserByID(duid)
         if fetchedUser == None:
             return jsonify(Error="No user found in the system with that id."), 404
-        
+
         mappedUser = self.mapUserToDict(fetchedUser)
-        return jsonify(User=mappedUser),200 #200 == OK
-        
+        return jsonify(User=mappedUser), 200  # 200 == OK
+
     def getDashUserByUsername(self, username):
         """
         Gets a single Dashboard user given their username.
@@ -130,17 +138,20 @@ class UserHandler:
 
         Args:
             username: The username of the dashboboard user that needs to be fetched.
-            
+
         Returns:
             A JSON containing all the user with the given username.
         """
+        if username == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
+
         dao = UserDAO()
         fetchedUser = dao.getDashUserByUsername(username)
         if fetchedUser == None:
             return jsonify(Error="No user found in the system with that username."), 404
-        
+
         mappedUser = self.mapUserToDict(fetchedUser)
-        return jsonify(User=mappedUser),200 #200 == OK
+        return jsonify(User=mappedUser), 200  # 200 == OK
 
     def getDashUserByEmail(self, email):
         """
@@ -152,17 +163,19 @@ class UserHandler:
 
         Args:
             email: The email of the dashboboard user that needs to be fetched.
-            
+
         Returns:
             A JSON containing all the user with the given email.
         """
+        if email == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
         dao = UserDAO()
         fetchedUser = dao.getDashUserByEmail(email)
         if fetchedUser == None:
             return jsonify(Error="No user found in the system with that email."), 404
-        
+
         mappedUser = self.mapUserToDict(fetchedUser)
-        return jsonify(User=mappedUser),200 #200 == OK
+        return jsonify(User=mappedUser), 200  # 200 == OK
 
     def login(self, username, password):
         """
@@ -178,30 +191,34 @@ class UserHandler:
         Returns:
             A JSON containing all the user's hash.
         """
+        if username == None or password == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
         dao = UserDAO()
 
-        #get user by username
-        user = dao.getDashUserByUsername(username) 
+        # get user by username
+        user = dao.getDashUserByUsername(username)
 
         # verify is user exists
         if user == None:
             # Login attempts do not increase because username does not exist.
             return jsonify(Error="Username or Password are incorrect."), 400
         # get user id , whic is position 0 of the user tupple.
-        duid = user[0] 
-        attempts = dao.getLoginAttempts(duid)[0] #get current number of attempts
-        
-        if attempts == 3 or user[4] == False : # If there are 3 failed login attempts, deactivate account.
+        duid = user[0]
+        # get current number of attempts
+        attempts = dao.getLoginAttempts(duid)[0]
+
+        # If there are 3 failed login attempts, deactivate account.
+        if attempts == 3 or user[4] == False:
             dao.deactivateDashUserAccount(duid)
-            return jsonify(Error='Account is locked, contact administrator.'), 401
+            return jsonify(Error='Account is locked.'), 401
 
         fetchedHash = dao.getHashByUsername(username)
-        if fetchedHash == None: #TODO this may be repeated code.
+        if fetchedHash == None:  # TODO this may be repeated code.
             # Login attempts do not increase because username does not exist.
             return jsonify(Error="Username or Password are incorrect."), 400
-        
+
         mappedHash = self.mapHash(fetchedHash)
-        #TODO AES encryption.
+        # TODO AES encryption.
         if verifyHash(password, mappedHash['hash']):
             # User provided correct password
             token = generateToken(username)
@@ -209,12 +226,12 @@ class UserHandler:
                 'user': username,
                 'token': token
             }
-            
-            dao.setLoginAttempts(duid,0)
+
+            dao.setLoginAttempts(duid, 0)
             session['username'] = loginInfo['user']
             return jsonify(auth=loginInfo), 201
 
-        dao.setLoginAttempts(duid,attempts+1)
+        dao.setLoginAttempts(duid, attempts+1)
         return jsonify(Error="Username or Password are incorrect."), 400
 
     def updateDashUserPassword(self, duid, password):
@@ -228,28 +245,81 @@ class UserHandler:
         Args:
             duid: The ID of the user whose password must be updated.
             password: The hash of the new password for the dashboboard user.
-            
+
         Returns:
             A JSON containing all the user with the updated dashboard user.
         """
+        if duid == None or password == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
+
         # Verify password complies with the rules
         if not rulesMatch(password):
             return jsonify(Error="""Password should contain At least 1 upercase letter,
             1 lowecase letter, at least 1 number, at least 1 symbol, and is between 
             10 and 64 characters long."""), 400
-        # Hash password 
+        # Hash password
         hashedPassword = createHash(password)
         dao = UserDAO()
         res = dao.updateDashUserPassword(duid, hashedPassword)
         if res == None:
             return jsonify(Error='No user found in the system with that id.'), 404
-        
-        ## When password is reset, login attempts are set to 0
-        dao.activateDashUserAccount(duid) #Set account active to true.
-        dao.setLoginAttempts(duid,0)
-        return jsonify(User=self.mapUserToDict(res)),201
 
-    def updateDashUserInfo(self, duid,username, full_name,email,is_active):
+        # When password is reset, login attempts are set to 0
+        dao.deactivateDashUserAccount(duid)  # Set account active to false.
+        dao.setLoginAttempts(duid, 0)
+        return jsonify(User=self.mapUserToDict(res)), 201
+
+    def unlockDashUserAccount(self, username, password, new_password):
+        """
+        Updates the password for the dashboard user with the given ID and sets is_active status.
+
+        Calls the UserDAO to update the password of a dashboard user and set its 
+        'is_active' status to true. It then maps the result to a JSON that contains 
+        the desired record. That JSON object is then returned.
+
+        Args:
+            username: The username of the user whose password must be updated.
+            password: Temporary password provided by the sistem admin.
+            new_password: New password set by the user.
+
+        Returns:
+            A JSON containing all the user with the updated dashboard user.
+        """
+        if username == None or password == None or new_password == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
+
+        dao = UserDAO()
+
+        #Verify if password matches the one on the database
+        fetchedHash = dao.getHashByUsername(username)
+        if fetchedHash == None: 
+            #username is incorrect 
+            return jsonify(Error="Username or Password are incorrect."), 400
+
+        mappedHash = self.mapHash(fetchedHash)
+        if not verifyHash(password, mappedHash['hash']):
+            #password is incorrect
+            return jsonify(Error="Username or Password are incorrect."), 400
+
+        # Verify the new password complies with the rules
+        if not rulesMatch(new_password):
+            return jsonify(Error="""Password should contain At least 1 upercase letter,
+            1 lowecase letter, at least 1 number, at least 1 symbol, and is between 
+            10 and 64 characters long."""), 400
+        
+        # If they do, hash password
+        hashedPassword = createHash(new_password)
+        
+        res = dao.updateDashUserPasswordByUsername(username, hashedPassword)
+        if res == None:
+            return jsonify(Error='No user found in the system with that id.'), 404
+        updatedUser = self.mapUserToDict(res)
+        # When password is reset, login attempts are set to 0
+        dao.activateDashUserAccount(updatedUser['id'])  # Set account active to true.
+        dao.setLoginAttempts(updatedUser['id'], 0)
+        return jsonify(User=updatedUser), 201
+
+    def updateDashUserInfo(self, duid, username, full_name, email, is_active):
         """
         Updates the username for the dashboard user with the given ID.
 
@@ -266,8 +336,12 @@ class UserHandler:
         Returns:
             A JSON containing all the user with the updated dashboard user. 
         """
+        if duid == None or username == None or full_name == None or email == None or is_active == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
+
         dao = UserDAO()
-        res = dao.updateDashUserInfo(duid, username, full_name,email,is_active)
+        res = dao.updateDashUserInfo(
+            duid, username, full_name, email, is_active)
         if res == None:
             return jsonify(Error='No user found in the system with that id.'), 404
         if res == 'UserError1':
@@ -275,9 +349,9 @@ class UserHandler:
         elif res == 'UserError2':
             return jsonify(Error='Username already taken.'), 400
         else:
-            return jsonify(User=self.mapUserToDict(res)),201
+            return jsonify(User=self.mapUserToDict(res)), 201
 
-    #TODO: Update documentation of the API merge mard inactive/active to toggle active
+    # TODO: Update documentation of the API merge mard inactive/active to toggle active
     def toggleDashUserActive(self, duid):
         """
         Toggle the active status of a dashboard user.
@@ -288,15 +362,17 @@ class UserHandler:
 
         Args:
             duid: The ID of the user that will be toggled.
-            
+
         Returns:
             A JSON containing the updated dashboard user.
         """
+        if duid == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
         dao = UserDAO()
         res = dao.toggleDashUserActive(duid)
         if res == None:
             return jsonify(Error='No user found in the system with that id.'), 404
-        return jsonify(User=self.mapUserToDict(res)),201
+        return jsonify(User=self.mapUserToDict(res)), 201
 
     def removeDashUser(self, duid):
         """
@@ -308,48 +384,54 @@ class UserHandler:
 
         Args:
             duid: The ID of the user that will be invalidated.
-            
+
         Returns:
             A JSON containing all the user with the updated dashboard user.
         """
+        if duid == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
         dao = UserDAO()
         res = dao.removeDashUser(duid)
         if res == None:
             return jsonify(Error='No user found in the system with that id.'), 404
-        return jsonify(User=self.mapUserToDict(res)),201
+        return jsonify(User=self.mapUserToDict(res)), 201
 
-    def setUserPermissions(self,duid,permissionsList):
-            """
-            Adds permissions to a user.
+    def setUserPermissions(self, duid, permissionsList):
+        """
+        Adds permissions to a user.
 
-            This fucntion will go thorugh the permissions list and apply them to 
-            the user with the specified duid.
+        This fucntion will go thorugh the permissions list and apply them to 
+        the user with the specified duid.
 
-            Args:
-                duid: The id of the user's whose permissions will be modified.
-                permissionsList: A list of the permissions to add to the user.
-            
-            Returns:
-                A list containing the response to the database query containing 
-                the matching record of modiffied user permissions.
-            """
-            for permission in permissionsList: #If at least one of the parameters of one the indexes is None, scrap the request
-                if 'permission_id' not in permission or 'is_invalid' not in permission:
-                    return jsonify(Error='Bad Request'), 400
+        Args:
+            duid: The id of the user's whose permissions will be modified.
+            permissionsList: A list of the permissions to add to the user.
 
-            dao = UserDAO()
-            resultList = dao.setUserPermissions(duid, permissionsList)
-            if resultList == 'UserError4':
-                return jsonify(Error='No User found in the system with that id.'), 404
-            if resultList == 'UserError5':
-                return jsonify(Error='Permissions cant be empty.'), 400 # Bad request
+        Returns:
+            A list containing the response to the database query containing 
+            the matching record of modiffied user permissions.
+        """
+        if duid == None or permissionsList == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
 
-            mappedPermissions = []
-            for row in resultList:
-                mappedPermissions.append(self.mapPermissionsToDict(row))
-            return jsonify(Permissions=mappedPermissions),201
+        for permission in permissionsList:  # If at least one of the parameters of one the indexes is None, scrap the request
+            if 'permission_id' not in permission or 'is_invalid' not in permission:
+                return jsonify(Error='Bad Request  '), 400
 
-    def getUserPermissions(self,duid):
+        dao = UserDAO()
+        resultList = dao.setUserPermissions(duid, permissionsList)
+        if resultList == 'UserError4':
+            return jsonify(Error='No User found in the system with that id.'), 404
+        if resultList == 'UserError5':
+            # Bad request
+            return jsonify(Error='Permissions cant be empty.'), 400
+
+        mappedPermissions = []
+        for row in resultList:
+            mappedPermissions.append(self.mapPermissionsToDict(row))
+        return jsonify(Permissions=mappedPermissions), 201
+
+    def getUserPermissions(self, duid):
         """
         Get permissions for a user.
 
@@ -361,6 +443,9 @@ class UserHandler:
         Returns:
             A list of permission dictionnaries ordered by permission id in ascending fashion.
         """
+        if duid == None:
+            return jsonify(Error="Request Parameters Undefined."), 400
+
         dao = UserDAO()
         permisionsList = dao.getUserPermissions(duid)
         if permisionsList == 'UserError4':
@@ -368,4 +453,4 @@ class UserHandler:
         mappedPermissions = []
         for row in permisionsList:
             mappedPermissions.append(self.mapPermissionsToDict(row))
-        return jsonify(Permissions=mappedPermissions),200
+        return jsonify(Permissions=mappedPermissions), 200
